Add style to "friedchicken.html"
Using CSS, I styled certain elements within the page. I also added
another image and resized both pictures to be smaller underneath
the "Fried Chicken" header.

Firstly, I used a universal selector to apply a font to all text
on the page. Secondly, I added and image sizing declaration as mentioned
earlier. Afterward, I created three classes; ".section-header",
".list" and ".ing" which were used throughout the page to style important
parts.

The sections titled "Description", "Ingredients", and "Steps" were given
a blue color and center-alignment. The ingredients list had two selectors
applied giving it a light green background and left alignment (this was
due to the fact that the .section-header rule make it centered and I
wanted it pushed to the left.)

Finally I added an ID selector just for the recipe credit. This styled
the text to be red, bold and with a different font.